Consolidate PromptAddView and PromptEditView into PromptFormView
Refactor two similar views into a unified PromptFormView that handles
both creating new prompts from templates and editing existing prompts.
Reduces code duplication and simplifies maintenance.
